@@ -14,93 +14,51 @@
   "license": "GPL-3.0",
   "scripts": {
     "clean": "rimraf dist/**/*",
-    "build": "npm run clean && tsc",
-    "build:watch": "npm run clean && tsc -watch",
-    "lint": "tslint src/**/*.ts spec/**/*.ts",
-    "lint:fix": "tslint src/**/*.ts spec/**/*.ts --fix",
+    "build": "npm run clean && ttsc",
+    "build:watch": "npm run clean && ttsc -watch",
+    "lint": "tslint '*/src/**/*.ts' 'spec/**/*.ts'",
+    "lint:fix": "tslint '*/src/**/*.ts' 'spec/**/*.ts' --fix",
     "test": "karma start ./spec/support/karma.conf.js --single-run",
     "test:watch": "karma start ./spec/support/karma.conf.js",
     "test:node": "npm run build && jasmine",
     "test:node:watch": "concurrently -k -n TSC,Node -c yellow,cyan \"npm run build:watch\" \"nodemon -w ./dist --delay 500ms --exec jasmine\""
   },
   "devDependencies": {
-    "@fluffy-spoon/substitute": "^1.179.0",
-    "@types/commander": "^2.12.2",
-    "@types/form-data": "^2.2.1",
-    "@types/inquirer": "^0.0.43",
-    "@types/jasmine": "^3.3.12",
-    "@types/lowdb": "^1.0.5",
-    "@types/lunr": "^2.3.3",
-    "@types/node": "^10.17.28",
-    "@types/node-fetch": "^2.1.2",
-    "@types/node-forge": "^0.7.5",
-    "@types/papaparse": "^4.5.3",
-    "@types/tldjs": "^2.3.0",
-    "@types/webcrypto": "0.0.28",
-    "@types/zxcvbn": "^4.4.0",
-    "concurrently": "3.5.1",
-    "cssstyle": "1.2.1",
-    "electron": "11.1.1",
-    "jasmine": "^3.3.1",
-    "jasmine-core": "^3.3.0",
-    "jasmine-spec-reporter": "^4.2.1",
+    "@fluffy-spoon/substitute": "^1.202.0",
+    "@types/jasmine": "^3.7.6",
+    "@types/node": "^14.17.1",
+    "concurrently": "^6.1.0",
+    "jasmine": "^3.7.0",
+    "jasmine-core": "^3.7.1",
     "jasmine-ts-console-reporter": "^3.1.1",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "^2.2.0",
+    "jsdom": "^16.5.3",
+    "karma": "^6.3.2",
+    "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-detect-browsers": "^2.3.3",
-    "karma-edge-launcher": "^0.4.2",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-firefox-launcher": "^2.1.0",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
     "karma-safari-launcher": "^1.0.0",
-    "karma-typescript": "^4.0.0",
-    "nodemon": "^1.17.3",
-    "rimraf": "^2.6.2",
-    "ts-node": "^9.1.0",
+    "karma-webpack": "^4.0.2",
+    "nodemon": "^2.0.7",
+    "rimraf": "^3.0.2",
+    "ts-loader": "^8.1.0",
     "tslint": "^6.1.3",
+    "ttypescript": "^1.5.12",
     "typemoq": "^2.1.0",
-    "typescript": "3.8.3"
+    "typescript": "4.1.5",
+    "typescript-transform-paths": "^2.2.3",
+    "webpack": "^4.46.0"
   },
   "dependencies": {
-    "@angular/animations": "9.1.12",
-    "@angular/cdk": "9.2.4",
-    "@angular/common": "9.1.12",
-    "@angular/compiler": "9.1.12",
-    "@angular/core": "9.1.12",
-    "@angular/forms": "9.1.12",
-    "@angular/platform-browser": "9.1.12",
-    "@angular/platform-browser-dynamic": "9.1.12",
-    "@angular/router": "9.1.12",
-    "@microsoft/signalr": "3.1.0",
-    "@microsoft/signalr-protocol-msgpack": "3.1.0",
-    "@nodert-win10-rs4/windows.security.credentials.ui": "^0.4.4",
-    "big-integer": "1.6.36",
-    "browser-hrtime": "^1.1.8",
-    "chalk": "2.4.1",
-    "commander": "2.18.0",
-    "core-js": "2.6.2",
-    "duo_web_sdk": "git+https://github.com/duosecurity/duo_web_sdk.git",
-    "electron-log": "4.3.0",
-    "electron-store": "6.0.1",
-    "electron-updater": "4.3.5",
-    "forcefocus": "^1.1.0",
-    "form-data": "2.3.2",
-    "https-proxy-agent": "5.0.0",
-    "inquirer": "6.2.0",
-    "jsdom": "13.2.0",
-    "keytar": "4.13.0",
-    "lowdb": "1.0.0",
-    "lunr": "2.3.3",
-    "ngx-infinite-scroll": "7.0.1",
-    "node-fetch": "2.2.0",
-    "node-forge": "0.7.6",
-    "open": "7.1.0",
-    "papaparse": "4.6.0",
-    "rxjs": "6.6.2",
-    "tldjs": "2.3.1",
-    "zone.js": "0.9.1",
-    "zxcvbn": "4.4.2"
+    "@bitwarden/jslib-angular": "file:angular",
+    "@bitwarden/jslib-common": "file:common",
+    "@bitwarden/jslib-electron": "file:electron",
+    "@bitwarden/jslib-node": "file:node"
+  },
+  "engines": {
+    "node": "~14",
+    "npm": "~7"
   }
 }
