@@ -1,14 +1,5 @@
-import { FieldType } from '../../../src/enums/fieldType';
-import { OnePassword1PifImporter as Importer } from '../../../src/importers/onepasswordImporters/onepassword1PifImporter';
-
-import { Utils } from '../../../src/misc/utils';
-
-if (Utils.isNode) {
-    // Polyfills
-    // tslint:disable-next-line
-    const jsdom: any = require('jsdom');
-    (global as any).DOMParser = new jsdom.JSDOM().window.DOMParser;
-}
+import { FieldType } from 'jslib-common/enums/fieldType';
+import { OnePassword1PifImporter as Importer } from 'jslib-common/importers/onepasswordImporters/onepassword1PifImporter';
 
 const TestData: string = '***aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee***\n' +
     JSON.stringify({
