@@ -1,17 +1,26 @@
-import { LastPassCsvImporter as Importer } from '../../../src/importers/lastpassCsvImporter';
+import { LastPassCsvImporter as Importer } from 'jslib-common/importers/lastpassCsvImporter';
 
-import { CipherView } from '../../../src/models/view/cipherView';
-import { FieldView } from '../../../src/models/view/fieldView';
+import { ImportResult } from 'jslib-common/models/domain/importResult';
+import { CipherView } from 'jslib-common/models/view/cipherView';
+import { FieldView } from 'jslib-common/models/view/fieldView';
 
-import { Utils } from '../../../src/misc/utils';
+import { CipherType, FieldType } from 'jslib-common/enums';
 
-import { FieldType } from '../../../src/enums';
+function baseExcept(result: ImportResult) {
+    expect(result).not.toBeNull();
+    expect(result.success).toBe(true);
+    expect(result.ciphers.length).toBe(1);
+}
 
-if (Utils.isNode) {
-    // Polyfills
-    // tslint:disable-next-line
-    const jsdom: any = require('jsdom');
-    (global as any).DOMParser = new jsdom.JSDOM().window.DOMParser;
+function expectLogin(cipher: CipherView) {
+    expect(cipher.type).toBe(CipherType.Login);
+
+    expect(cipher.name).toBe('example.com');
+    expect(cipher.notes).toBe('super secure notes');
+    expect(cipher.login.uri).toBe('http://example.com');
+    expect(cipher.login.username).toBe('someUser');
+    expect(cipher.login.password).toBe('myPassword');
+    expect(cipher.login.totp).toBe('Y64VEVMBTSXCYIWRSHRNDZW62MPGVU2G');
 }
 
 const CipherData = [
@@ -160,7 +169,7 @@ Notes:",nomonth,,0`,
 ];
 
 describe('Lastpass CSV Importer', () => {
-    CipherData.forEach((data) => {
+    CipherData.forEach(data => {
         it(data.title, async () => {
             const importer = new Importer();
             const result = await importer.parse(data.csv);
@@ -176,5 +185,17 @@ describe('Lastpass CSV Importer', () => {
                 }
             }
         });
+    });
+
+    it('should parse login with totp', async () => {
+        const input = `url,username,password,totp,extra,name,grouping,fav
+        http://example.com,someUser,myPassword,Y64VEVMBTSXCYIWRSHRNDZW62MPGVU2G,super secure notes,example.com,,0`;
+
+        const importer = new Importer();
+        const result = await importer.parse(input);
+        baseExcept(result);
+
+        const cipher = result.ciphers[0];
+        expectLogin(cipher);
     });
 });
