@@ -1,4 +1,4 @@
-import { Utils } from '../../../src/misc/utils';
+import { Utils } from 'jslib-common/misc/utils';
 
 describe('Utils Service', () => {
     describe('getDomain', () => {
@@ -32,6 +32,11 @@ describe('Utils Service', () => {
         it('should support localhost and IP', () => {
             expect(Utils.getDomain('https://localhost')).toBe('localhost');
             expect(Utils.getDomain('https://192.168.1.1')).toBe('192.168.1.1');
+        });
+
+        it('should reject invalid hostnames', () => {
+            expect(Utils.getDomain('https://mywebsite.com$.mywebsite.com')).toBeNull();
+            expect(Utils.getDomain('https://mywebsite.com!.mywebsite.com')).toBeNull();
         });
     });
 
